[server] Support `canopy stop` on Windows
Two bugs:
1. Trying to `pgrep gunicorn` on Windows, even though windows doesn't have `pgrep` and Gunicorn isn't supported on windows.
2. In windows we need to send the `CTRL_C` signal, which has a totally different implmentation behind it than GITINT

+++ src/canopy_cli/cli.py
@@ -577,28 +577,29 @@ def start(host: str, port: str, reload: bool,
 @click.option("url", "--url", default="http://0.0.0.0:8000",
               help="URL of the Canopy server to use. Defaults to http://0.0.0.0:8000")
 def stop(url):
-    # Check if the server was started using Gunicorn
-    res = subprocess.run(["pgrep", "-f", "gunicorn canopy_server.app:app"],
-                         capture_output=True)
-    output = res.stdout.decode("utf-8").split()
+    if os.name != "nt":
+        # Check if the server was started using Gunicorn
+        res = subprocess.run(["pgrep", "-f", "gunicorn canopy_server.app:app"],
+                             capture_output=True)
+        output = res.stdout.decode("utf-8").split()
 
-    # If Gunicorn was used, kill all Gunicorn processes
-    if output:
-        msg = ("It seems that Canopy server was launched using Gunicorn.\n"
-               "Do you want to kill all Gunicorn processes?")
-        click.confirm(click.style(msg, fg="red"), abort=True)
-        try:
-            subprocess.run(["pkill", "-f", "gunicorn canopy_server.app:app"],
-                           check=True)
-        except subprocess.CalledProcessError:
+        # If Gunicorn was used, kill all Gunicorn processes
+        if output:
+            msg = ("It seems that Canopy server was launched using Gunicorn.\n"
+                   "Do you want to kill all Gunicorn processes?")
+            click.confirm(click.style(msg, fg="red"), abort=True)
             try:
-                [os.kill(int(pid), signal.SIGINT) for pid in output]
-            except OSError:
-                msg = (
-                    "Could not kill Gunicorn processes. Please kill them manually."
-                    f"Found process ids: {output}"
-                )
-                raise CLIError(msg)
+                subprocess.run(["pkill", "-f", "gunicorn canopy_server.app:app"],
+                               check=True)
+            except subprocess.CalledProcessError:
+                try:
+                    [os.kill(int(pid), signal.SIGINT) for pid in output]
+                except OSError:
+                    msg = (
+                        "Could not kill Gunicorn processes. Please kill them manually."
+                        f"Found process ids: {output}"
+                    )
+                    raise CLIError(msg)
 
     try:
         res = requests.get(urljoin(url, "/shutdown"))
